PE finder moved to superclass
