+++ resources/CommandLineQAList.txt
@@ -32,4 +32,4 @@ org.gk.qualityCheck.NonHumanEventsNotManuallyInferredCheck
 org.gk.qualityCheck.NonHumanPhysicalEntitiesWithoutDiseaseCheck
 org.gk.qualityCheck.NonHumanReactionsWithHumanPhysicalEntitiesCheck
 org.gk.qualityCheck.HumanReactionsWithoutDiseaseAndHaveNonHumanPhysicalEntities
-org.gk.qualityCheck.HumanReactionWithNonHumanComplexWithHumanComponentCheck
+org.gk.qualityCheck.HumanReactionsWithNonHumanComplexWithHumanComponentCheck
