+++ src/org/gk/qualityCheck/NonHumanPhysicalEntitiesWithoutDiseaseCheck.java
@@ -12,7 +12,6 @@ import java.util.*;
 public class NonHumanPhysicalEntitiesWithoutDiseaseCheck extends NonHumanEventsNotManuallyInferredCheck {
 
     private static List<String> skiplistDbIDs = new ArrayList<>();
-    private static GKInstance humanSpeciesInst = new GKInstance();
 
     @Override
     public QAReport checkInCommand() throws Exception {
@@ -21,7 +20,6 @@ public class NonHumanPhysicalEntitiesWithoutDiseaseCheck extends NonHumanEventsN
             return null;
         }
         MySQLAdaptor dba = (MySQLAdaptor) dataSource;
-        humanSpeciesInst = dba.fetchInstance(48887L);
         List<GKInstance> reactionsNotUsedForManualInference = findReactionsNotUsedForManualInference(dba);
         for (GKInstance reaction : reactionsNotUsedForManualInference) {
             if (reaction.getSchemClass().isa(ReactomeJavaConstants.ReactionlikeEvent)) {
@@ -76,134 +74,6 @@ public class NonHumanPhysicalEntitiesWithoutDiseaseCheck extends NonHumanEventsN
 
     private void setSkipList() throws IOException {
         skiplistDbIDs = Files.readAllLines(Paths.get("QA_SkipList/Manually_Curated_NonHuman_Pathways.txt"));
-    }
-
-    private boolean hasNonHumanSpecies(GKInstance reactionPE) throws Exception {
-        if (reactionPE.getSchemClass().isa(ReactomeJavaConstants.Drug)) {
-            return false;
-        }
-        if (reactionPE.getAttributeValue(ReactomeJavaConstants.species) == null) {
-            return false;
-        }
-        if (reactionPE.getAttributeValue(ReactomeJavaConstants.species) == humanSpeciesInst) {
-            return false;
-        }
-
-        return true;
-    }
-
-    private boolean hasSpeciesAttribute(GKInstance physicalEntity) {
-        if (physicalEntity.getSchemClass().isa(ReactomeJavaConstants.OtherEntity) ||
-                physicalEntity.getSchemClass().isa(ReactomeJavaConstants.Drug) ||
-                physicalEntity.getSchemClass().isa(ReactomeJavaConstants.SimpleEntity)) {
-            return false;
-        }
-        return true;
-    }
-
-    private Set<GKInstance> findAllPhysicalEntitiesInReaction(GKInstance reaction) throws Exception {
-        Set<GKInstance> reactionPEs = new HashSet<>();
-        reactionPEs.addAll(findAllInputAndOutputPEs(reaction));
-        reactionPEs.addAll(findAllCatalystPEs(reaction));
-        reactionPEs.addAll(findAllRegulationPEs(reaction));
-        return reactionPEs;
-    }
-
-    private Set<GKInstance> findAllInputAndOutputPEs(GKInstance reaction) throws Exception {
-        Set<GKInstance> inputOutputPEs = new HashSet<>();
-        for (String attribute : Arrays.asList(ReactomeJavaConstants.input, ReactomeJavaConstants.output)) {
-            for (GKInstance attributePE : (Collection<GKInstance>) reaction.getAttributeValuesList(attribute)) {
-                if (hasSpeciesAttribute(attributePE)) {
-                    inputOutputPEs.add(attributePE);
-                    inputOutputPEs.addAll(findAllPhysicalEntities(attributePE));
-                }
-            }
-        }
-        return inputOutputPEs;
-    }
-
-    private Set<GKInstance> findAllCatalystPEs(GKInstance reaction) throws Exception {
-        Set<GKInstance> catalystPEs = new HashSet<>();
-        List<GKInstance> catalysts = reaction.getAttributeValuesList(ReactomeJavaConstants.catalystActivity);
-        for (GKInstance catalyst : catalysts) {
-            for (String attribute : Arrays.asList(ReactomeJavaConstants.activeUnit, ReactomeJavaConstants.physicalEntity)) {
-                for (GKInstance attributePE : (Collection<GKInstance>) catalyst.getAttributeValuesList(attribute)) {
-                    if (hasSpeciesAttribute(attributePE)) {
-                        catalystPEs.add(attributePE);
-                        catalystPEs.addAll(findAllPhysicalEntities(attributePE));
-                    }
-                }
-            }
-        }
-        return catalystPEs;
-    }
-
-    private Set<GKInstance> findAllRegulationPEs(GKInstance reaction) throws Exception {
-        Set<GKInstance> regulationPEs = new HashSet<>();
-        List<GKInstance> regulations = reaction.getAttributeValuesList(ReactomeJavaConstants.regulatedBy);
-        for (GKInstance regulation : regulations) {
-            for (String attribute : Arrays.asList(ReactomeJavaConstants.activeUnit, ReactomeJavaConstants.regulator)) {
-                for (GKInstance attributePE : (Collection<GKInstance>) regulation.getAttributeValuesList(attribute)) {
-                    if (hasSpeciesAttribute(attributePE)) {
-                        regulationPEs.add(attributePE);
-                        regulationPEs.addAll(findAllPhysicalEntities(attributePE));
-                    }
-                }
-            }
-        }
-        return regulationPEs;
-    }
-
-    private Set<GKInstance> findAllPhysicalEntities(GKInstance physicalEntity) throws Exception {
-        Set<GKInstance> physicalEntities = new HashSet<>();
-        if (hasSpeciesAttribute(physicalEntity)) {
-            if (containsMultiplePEs(physicalEntity)) {
-                physicalEntities.add(physicalEntity);
-                physicalEntities.addAll(findAllConstituentPEs(physicalEntity));
-            } else if (physicalEntity.getSchemClass().isa(ReactomeJavaConstants.EntityWithAccessionedSequence)) {
-                physicalEntities.add(physicalEntity);
-            }
-        }
-        return physicalEntities;
-    }
-
-    private boolean containsMultiplePEs(GKInstance physicalEntity) {
-        if (physicalEntity.getSchemClass().isa(ReactomeJavaConstants.Complex) ||
-                physicalEntity.getSchemClass().isa(ReactomeJavaConstants.Polymer) ||
-                physicalEntity.getSchemClass().isa(ReactomeJavaConstants.EntitySet)) {
-            return true;
-        }
-        return false;
-    }
-
-    private Set<GKInstance> findAllConstituentPEs(GKInstance physicalEntity) throws Exception {
-        Set<GKInstance> physicalEntities = new HashSet<>();
-        if (physicalEntity.getSchemClass().isa(ReactomeJavaConstants.Complex)) {
-            for (GKInstance complexComponent : (Collection<GKInstance>) physicalEntity.getAttributeValuesList(ReactomeJavaConstants.hasComponent)) {
-                physicalEntities.addAll(findAllPhysicalEntities(complexComponent));
-            }
-        } else if (physicalEntity.getSchemClass().isa(ReactomeJavaConstants.Polymer)) {
-            for (GKInstance polymerUnit : (Collection<GKInstance>) physicalEntity.getAttributeValuesList(ReactomeJavaConstants.repeatedUnit)) {
-                physicalEntities.addAll(findAllPhysicalEntities(polymerUnit));
-            }
-        } else if (physicalEntity.getSchemClass().isa(ReactomeJavaConstants.EntitySet)) {
-            physicalEntities.addAll(findEntitySetPhysicalEntities(physicalEntity));
-        }
-        return physicalEntities;
-    }
-
-    private Set<GKInstance> findEntitySetPhysicalEntities(GKInstance physicalEntity) throws Exception {
-        Set<GKInstance> physicalEntities = new HashSet<>();
-        Set<String> entitySetAttributes = new HashSet<>(Arrays.asList(ReactomeJavaConstants.hasMember, ReactomeJavaConstants.hasCandidate));
-        if (physicalEntity.getSchemClass().isa(ReactomeJavaConstants.DefinedSet)) {
-            entitySetAttributes.remove(ReactomeJavaConstants.hasCandidate);
-        }
-        for (String entitySetAttribute : entitySetAttributes) {
-            for (GKInstance setInstance : (Collection<GKInstance>) physicalEntity.getAttributeValuesList(entitySetAttribute)) {
-                physicalEntities.addAll(findAllPhysicalEntities((setInstance)));
-            }
-        }
-        return physicalEntities;
     }
 
     private String getReportLine(GKInstance reactionPE, GKInstance reaction) throws Exception {
